Solutions for the JS-Front-End course at SoftUni
I've added all the things I have written during the JS-Front-End course at SoftUni. I plan to fill the missing task in the coming days if any fellow programmer has difficulty with any of the tasks. Note - the course difficulty is supposed to be at fundamental level, so no advanced topics are to be expected.